requirements.txt
Django
django-crispy-forms
djangorestframework
django_filter